+++ R/get_data.R
@@ -345,7 +345,7 @@ get_pop_data <- function(data){
   data_pop <- data %>%
     mutate(Country.Region = country_name) %>%
     left_join(population) %>%
-    filter(!is.na(population))
+    filter(!is.na(population) & population > 1000)
   # select(-Country.Region)
 
   data_pop