+++ R/mod_global.R
@@ -41,6 +41,9 @@ mod_global_ui <- function(id){
     fluidRow(
       column(12,offset = 6,
              mod_scatterplot_ui(ns("plot_scatterplot_glob"))
+             ),
+      column(12,
+             mod_stackedbarplot_ui(ns("plot_stackedbarplot_status"))
       )
     ),
     hr(),
@@ -146,6 +149,9 @@ mod_global_server <- function(input, output, session, orig_data, orig_data_aggre
 
   # > scatterplot prevalence vs growth
   callModule(mod_scatterplot_server, "plot_scatterplot_glob", orig_data_aggregate)
+  
+  # > stacked barplot with status split
+  callModule(mod_stackedbarplot_status_server, "plot_stackedbarplot_status", orig_data_aggregate)
 
   # tables ----
   callModule(mod_add_table_server, "add_table_world", world)