+++ NEWS.md
@@ -1,11 +1,15 @@
 ### Covid19 1.1.6.9000 (development version)
 
+- set different color -grey for NA in new active mod_map.R (#59)
+- fix colour scale/legend over 1M by replacing data_clean() with data_plot() (#79)
+- Country comp Lineplot from day of 1000 contagion rolling weekly average, remove dots, add imports zoo (#80)
 - Fix colour order in mod_lineplots_day_contagion (#67)
 - Merge the two boxes for current and new values into one.(#69)
 - fix trend order of area plot both global and country tabs, added utils funtion (#74)
 - correct plot titles for modules mod_compare_nth_cases_plot and mod_growth_death_rate (#66)
 - Add angle to bar-plot labels (#68)
 - Add scatterplot prevalence vs growth (#81)
+- Add status split stackedbarplot (#83)
 
 ### Covid19 1.1.6 (2020-04-06)